Fix UBSan left-shift UB and clang-tidy warnings
- chunk_coords.hpp: replace (chunk_axis << 5) with * CHUNK_WHICH constant;
  left-shifting a negative int32 is UB (ASan+UBSan leg: 'left shift of
  negative value -1'). Multiplication is bit-identical for valid inputs.
- player_controller.cpp: merge duplicated branch bodies (bugprone-branch-clone);
  sneaking/release both set sprint_active_ = false in one guard.
- player_controller.hpp: NOLINT GDCLASS operator= (bugprone-unhandled-self-assignment).
- mesh_manager.cpp: remove dead dx/dy/dz distance-to-chunk-bounds stores
  (clang-analyzer-deadcode.DeadStores) plus the now-unused min/max helpers.
- mesh_types.hpp: size_t-cast CHUNK_WIDTH*CHUNK_DEPTH array extent
  (bugprone-implicit-widening-of-multiplication-result).

+++ src/mesh/mesh_types.hpp
@@ -1,6 +1,7 @@
 #ifndef FUK_MINECRAFT_MESH_TYPES_HPP
 #define FUK_MINECRAFT_MESH_TYPES_HPP
 #include <cstdint>
+#include <cstddef>
 #include <array>
 #include <vector>
 #include "core/chunk_coords.hpp"
@@ -71,7 +72,9 @@ struct CachedQuad {
 // against the previous build's values to find every column whose light changed
 // since then, so light-only changes are re-emitted too (not just block edits).
 struct MeshLightChecksums {
-    std::array<uint32_t, CHUNK_WIDTH * CHUNK_DEPTH> columns{};
+    std::array<uint32_t,
+               static_cast<std::size_t>(CHUNK_WIDTH) * static_cast<std::size_t>(CHUNK_DEPTH)>
+        columns{};
 };
 
 #ifndef FUZZING_BUILD_MODE_UNSAFE_FOR_PRODUCTION
